[Refactor] change images
